requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alluka==0.1.4
attrs==23.2.0
colorlog==6.8.2
frozenlist==1.4.1
hikari==2.0.0.dev122
hikari-arc==1.2.1
hikari-miru==4.0.0
idna==3.6
multidict==6.0.5
yarl==1.9.4
